WEBNEW-180 :white_check_mark: Updated filter label tests to be rtl
